+++ Editors/Audio/BnkCompiler/ObjectGeneration/Warhammer3/DialogueEventGenerator.cs
@@ -31,9 +31,9 @@ namespace Editors.Audio.BnkCompiler.ObjectGeneration.Warhammer3
             return ConvertToWwise(typedProjectItem, project);
         }
 
-        public CAkDialogueEvent_V136TEMP ConvertToWwise(DialogueEvent inputDialogueEvent, CompilerData project)
+        public CAkDialogueEvent_V136 ConvertToWwise(DialogueEvent inputDialogueEvent, CompilerData project)
         {
-            var wwiseDialogueEvent = new CAkDialogueEvent_V136TEMP();
+            var wwiseDialogueEvent = new CAkDialogueEvent_V136();
             wwiseDialogueEvent.Arguments = new List<AkGameSync_V136>();
             wwiseDialogueEvent.Id = inputDialogueEvent.Id;
             wwiseDialogueEvent.HircType = AkBkHircType.Dialogue_Event;
