+++ Shared/GameFiles/Wwise/Hirc/HircFactory.cs
@@ -37,35 +37,35 @@ namespace Shared.GameFormats.Wwise.Hirc
         private static HircFactory CreateFactory_v112()
         {
             var instance = new HircFactory();
-            instance.RegisterHirc(AkBkHircType.ActorMixer, () => new V112.CAkActorMixer_V112TEMP());
+            instance.RegisterHirc(AkBkHircType.ActorMixer, () => new V112.CAkActorMixer_V112());
             instance.RegisterHirc(AkBkHircType.Sound, () => new V112.CAkSound_V112());
-            instance.RegisterHirc(AkBkHircType.Event, () => new V112.CAkEvent_V112TEMP());
-            instance.RegisterHirc(AkBkHircType.Action, () => new V112.CAkAction_V112TEMP());
+            instance.RegisterHirc(AkBkHircType.Event, () => new V112.CAkEvent_V112());
+            instance.RegisterHirc(AkBkHircType.Action, () => new V112.CAkAction_V112());
             instance.RegisterHirc(AkBkHircType.SwitchContainer, () => new V112.CAkSwitchCntr_V112());
             instance.RegisterHirc(AkBkHircType.SequenceContainer, () => new V112.CAkRanSeqCnt_V112());
-            instance.RegisterHirc(AkBkHircType.LayerContainer, () => new V112.CAkLayerCntr_V112TEMP());
-            instance.RegisterHirc(AkBkHircType.Dialogue_Event, () => new V112.CAkDialogueEvent_V112TEMP());
+            instance.RegisterHirc(AkBkHircType.LayerContainer, () => new V112.CAkLayerCntr_V112());
+            instance.RegisterHirc(AkBkHircType.Dialogue_Event, () => new V112.CAkDialogueEvent_V112());
             return instance;
         }
 
         private static HircFactory CreateFactory_v136()
         {
             var instance = new HircFactory();
-            instance.RegisterHirc(AkBkHircType.ActorMixer, () => new V136.CAkActorMixer_V136TEMP());
-            instance.RegisterHirc(AkBkHircType.Sound, () => new V136.CAkSound_V136TEMP());
-            instance.RegisterHirc(AkBkHircType.Event, () => new V136.CAkEvent_V136TEMP());
-            instance.RegisterHirc(AkBkHircType.Action, () => new V136.CAkAction_V136TEMP());
-            instance.RegisterHirc(AkBkHircType.SwitchContainer, () => new V136.CAkSwitchCntr_V136TEMP());
-            instance.RegisterHirc(AkBkHircType.SequenceContainer, () => new V136.CAkRanSeqCntr_V136TEMP());
-            instance.RegisterHirc(AkBkHircType.LayerContainer, () => new V136.CAkLayerCntr_V136TEMP());
-            instance.RegisterHirc(AkBkHircType.Dialogue_Event, () => new V136.CAkDialogueEvent_V136TEMP());
-            instance.RegisterHirc(AkBkHircType.Music_Track, () => new V136.CAkMusicTrack_V136TEMP());
-            instance.RegisterHirc(AkBkHircType.Music_Segment, () => new V136.CAkMusicSegment_V136TEMP());
-            instance.RegisterHirc(AkBkHircType.Music_Random_Sequence, () => new V136.CAkMusicRanSeqCntr_V136TEMP());
-            instance.RegisterHirc(AkBkHircType.Music_Switch, () => new V136.CAkMusicSwitchCntr_V136TEMP());
-            instance.RegisterHirc(AkBkHircType.FxCustom, () => new V136.CAkFxCustom_V136TEMP());
-            instance.RegisterHirc(AkBkHircType.FxShareSet, () => new V136.CAkFxShareSet_V136TEMP());
-            instance.RegisterHirc(AkBkHircType.Audio_Bus, () => new V136.CAkBus_V136TEMP());
+            instance.RegisterHirc(AkBkHircType.ActorMixer, () => new V136.CAkActorMixer_V136());
+            instance.RegisterHirc(AkBkHircType.Sound, () => new V136.CAkSound_V136());
+            instance.RegisterHirc(AkBkHircType.Event, () => new V136.CAkEvent_V136());
+            instance.RegisterHirc(AkBkHircType.Action, () => new V136.CAkAction_V136());
+            instance.RegisterHirc(AkBkHircType.SwitchContainer, () => new V136.CAkSwitchCntr_V136());
+            instance.RegisterHirc(AkBkHircType.SequenceContainer, () => new V136.CAkRanSeqCntr_V136());
+            instance.RegisterHirc(AkBkHircType.LayerContainer, () => new V136.CAkLayerCntr_V136());
+            instance.RegisterHirc(AkBkHircType.Dialogue_Event, () => new V136.CAkDialogueEvent_V136());
+            instance.RegisterHirc(AkBkHircType.Music_Track, () => new V136.CAkMusicTrack_V136());
+            instance.RegisterHirc(AkBkHircType.Music_Segment, () => new V136.CAkMusicSegment_V136());
+            instance.RegisterHirc(AkBkHircType.Music_Random_Sequence, () => new V136.CAkMusicRanSeqCntr_V136());
+            instance.RegisterHirc(AkBkHircType.Music_Switch, () => new V136.CAkMusicSwitchCntr_V136());
+            instance.RegisterHirc(AkBkHircType.FxCustom, () => new V136.CAkFxCustom_V136());
+            instance.RegisterHirc(AkBkHircType.FxShareSet, () => new V136.CAkFxShareSet_V136());
+            instance.RegisterHirc(AkBkHircType.Audio_Bus, () => new V136.CAkBus_V136());
             return instance;
         }
     }
