+++ Shared/GameFiles/Wwise/Hirc/V112/CAkLayerCntr_V112.cs
@@ -1,15 +1,16 @@
 ﻿using Shared.Core.ByteParsing;
 using Shared.GameFormats.Wwise.Enums;
+using Shared.GameFormats.Wwise.Hirc.V112.Shared;
 using Shared.GameFormats.Wwise.Hirc.V136.Shared;
 
-namespace Shared.GameFormats.Wwise.Hirc.V136
+namespace Shared.GameFormats.Wwise.Hirc.V112
 {
-    public class CAkLayerCntr_V136TEMP : HircItem, ICAkLayerCntr
+    public class CAkLayerCntr_V112 : HircItem, ICAkLayerCntr
     {
-        public NodeBaseParams_V136 NodeBaseParams { get; set; } = new NodeBaseParams_V136();
-        public Children_V136 Children { get; set; } = new Children_V136();
+        public NodeBaseParams_V112 NodeBaseParams { get; set; } = new NodeBaseParams_V112();
+        public Children_V112 Children { get; set; } = new Children_V112();
         public uint NumLayers { get; set; }
-        public List<CAkLayer_V136> LayerList { get; set; } = [];
+        public List<CAkLayer_V112> LayerList { get; set; } = [];
         public byte IsContinuousValidation { get; set; }
 
         protected override void CreateSpecificData(ByteChunk chunk)
@@ -20,12 +21,10 @@ namespace Shared.GameFormats.Wwise.Hirc.V136
             NumLayers = chunk.ReadUInt32();
             for (var i = 0; i < NumLayers; i++)
             {
-                var layer = new CAkLayer_V136();
+                var layer = new CAkLayer_V112();
                 layer.Create(chunk);
                 LayerList.Add(layer);
             }
-
-            IsContinuousValidation = chunk.ReadByte();
         }
 
         public override byte[] GetAsByteArray() => throw new NotSupportedException("Users probably don't need this complexity.");
@@ -34,18 +33,18 @@ namespace Shared.GameFormats.Wwise.Hirc.V136
         public List<uint> GetChildren() => Children.ChildIds;
         public uint GetDirectParentId() => NodeBaseParams.DirectParentId;
 
-        public class CAkLayer_V136
+        public class CAkLayer_V112
         {
-            public uint LayerId { get; set; }
-            public InitialRtpc_V136 InitialRtpc { get; set; } = new InitialRtpc_V136();
+            public uint UlLayerIr { get; set; }
+            public InitialRtpc_V112 InitialRtpc { get; set; } = new InitialRtpc_V112();
             public uint RtpcId { get; set; }
             public AkRtpcType RtpcType { get; set; }
-            public uint NumAssoc { get; set; }
-            public List<CAssociatedChildData_V136> CAssociatedChildDataList { get; set; } = [];
+            public uint NumAssoc {  get; set; }
+            public List<CAssociatedChildData_V112> CAssociatedChildDataList { get; set; } = [];
 
             public void Create(ByteChunk chunk)
             {
-                LayerId = chunk.ReadUInt32();
+                UlLayerIr = chunk.ReadUInt32();
                 InitialRtpc.Create(chunk);
                 RtpcId = chunk.ReadUInt32();
                 RtpcType = (AkRtpcType)chunk.ReadByte();
@@ -53,29 +52,25 @@ namespace Shared.GameFormats.Wwise.Hirc.V136
                 NumAssoc = chunk.ReadUInt32();
                 for (var i = 0; i < NumAssoc; i++)
                 {
-                    var associatedChildData = new CAssociatedChildData_V136();
+                    var associatedChildData = new CAssociatedChildData_V112();
                     associatedChildData.Create(chunk);
                     CAssociatedChildDataList.Add(associatedChildData);
                 }
             }
         }
 
-        public class CAssociatedChildData_V136
+        public class CAssociatedChildData_V112
         {
             public uint AssociatedChildId { get; set; }
-            public byte UnknownCustom0 { get; set; }
-            public byte UnknownCustom1 { get; set; }
-            public uint CurveSize {  get; set; }
-            public List<AkRtpcGraphPoint_V136> AkRtpcGraphPointList { get; set; } = [];
+            public uint CurveSize { get; set; }
+            public List<AkRtpcGraphPoint_V112> AkRtpcGraphPointList { get; set; } = [];
 
             public void Create(ByteChunk chunk)
             {
                 AssociatedChildId = chunk.ReadUInt32();
-                UnknownCustom0 = chunk.ReadByte();
-                UnknownCustom1 = chunk.ReadByte();
                 CurveSize = chunk.ReadUInt32();
                 for (var i = 0; i < CurveSize; i++)
-                    AkRtpcGraphPointList.Add(AkRtpcGraphPoint_V136.Create(chunk));
+                    AkRtpcGraphPointList.Add(AkRtpcGraphPoint_V112.Create(chunk));
             }
         }
     }
