+++ Shared/GameFiles/Wwise/Hirc/V136/CAkBus_V136.cs
@@ -3,7 +3,7 @@ using Shared.GameFormats.Wwise.Hirc.V136.Shared;
 
 namespace Shared.GameFormats.Wwise.Hirc.V136
 {
-    public class CAkBus_V136TEMP : HircItem
+    public class CAkBus_V136 : HircItem
     {
         public uint OverrideBusId { get; set; }
         public uint IdDeviceShareset { get; set; }
