+++ Shared/GameFiles/Wwise/Hirc/V136/CAkFxShareSet_V136.cs
@@ -3,7 +3,7 @@ using Shared.GameFormats.Wwise.Hirc.V136.Shared;
 
 namespace Shared.GameFormats.Wwise.Hirc.V136
 {
-    public class CAkFxCustom_V136TEMP : HircItem
+    public class CAkFxShareSet_V136 : HircItem
     {
         public uint PluginId { get; set; }
         public uint Size { get; set; }
@@ -12,7 +12,7 @@ namespace Shared.GameFormats.Wwise.Hirc.V136
         public List<AkMediaMap_V136> MediaList { get; set; } = [];
         public InitialRtpc_V136 InitialRtpc { get; set; } = new InitialRtpc_V136();
         public StateChunk_V136 StateChunk { get; set; } = new StateChunk_V136();
-        public short NumValues { get; set; }
+        public ushort NumValues { get; set; }
         public List<PluginPropertyValue_V136> PropertyValuesList { get; set; } = [];
 
         protected override void CreateSpecificData(ByteChunk chunk)
@@ -28,7 +28,7 @@ namespace Shared.GameFormats.Wwise.Hirc.V136
             InitialRtpc.Create(chunk);
             StateChunk.Create(chunk);
 
-            NumValues = chunk.ReadShort();
+            NumValues = chunk.ReadUShort();
             for (var i = 0; i < NumValues; i++)
                 PropertyValuesList.Add(PluginPropertyValue_V136.Create(chunk));
         }
