+++ Shared/GameFiles/Wwise/Hirc/V136/CAkMusicRanSeqCntr_V136.cs
@@ -3,7 +3,7 @@ using Shared.GameFormats.Wwise.Hirc.V136.Shared;
 
 namespace Shared.GameFormats.Wwise.Hirc.V136
 {
-    public class CAkMusicRanSeqCntr_V136TEMP : HircItem
+    public class CAkMusicRanSeqCntr_V136 : HircItem
     {
         public MusicTransNodeParams_V136 MusicTransNodeParams { get; set; } = new MusicTransNodeParams_V136();
         public uint NumPlaylistItems { get; set; }
