+++ Shared/GameFiles/Wwise/Hirc/V136/CAkMusicSegment_V136.cs
@@ -4,7 +4,7 @@ using Shared.GameFormats.Wwise.Hirc.V136.Shared;
 
 namespace Shared.GameFormats.Wwise.Hirc.V136
 {
-    public partial class CAkMusicSegment_V136TEMP : HircItem
+    public partial class CAkMusicSegment_V136 : HircItem
     {
         public MusicNodeParams_V136 MusicNodeParams { get; set; } = new MusicNodeParams_V136();
         public double Duration { get; set; }
